Split SourcesPage into more files
