+++ src/pages/SourcesPage.js
@@ -6,100 +6,30 @@ import {
     loadAppTypes,
     loadEntities,
     loadSourceTypes,
-    clearAddSource,
     filterProviders
 } from '../redux/actions/providers';
 import { Button } from '@patternfly/react-core';
 import { FormattedMessage, useIntl } from 'react-intl';
 import { AddSourceWizard } from '@redhat-cloud-services/frontend-components-sources';
 import { PrimaryToolbar } from '@redhat-cloud-services/frontend-components';
-import awesomeDebounce from 'awesome-debounce-promise';
 
 import SourcesSimpleView from '../components/SourcesSimpleView/SourcesSimpleView';
 import SourcesEmptyState from '../components/SourcesEmptyState';
 import SourceEditModal from '../components/SourceEditForm/SourceEditModal';
 import SourceRemoveModal from '../components/SourceRemoveModal';
 import AddApplication from '../components/AddApplication/AddApplication';
-import { pageAndSize, addMessage } from '../redux/actions/providers';
+import { pageAndSize } from '../redux/actions/providers';
 import { paths } from '../Routes';
-import UndoButtonAdd from '../components/UndoButton/UndoButtonAdd';
-import isEmpty from 'lodash/isEmpty';
 
-export const onCloseAddSourceWizard = ({ values, dispatch, history, intl }) => {
-    if (values && !isEmpty(values)) {
-        const messageId = Date.now();
-        dispatch(addMessage(
-            intl.formatMessage({
-                id: 'sources.addWizardCanceled',
-                defaultMessage: 'Adding a source was cancelled'
-            }),
-            'success',
-            <FormattedMessage
-                id="sources.undoMistake"
-                defaultMessage={ `{undo} if this was a mistake.` }
-                values={ { undo: <UndoButtonAdd messageId={messageId} values={values} /> } }
-            />,
-            messageId
-        ));
-    }
-
-    dispatch(clearAddSource());
-    history.push('/');
-};
-
-export const debouncedFiltering = awesomeDebounce((refresh) => refresh(), 500);
-
-export const afterSuccessLoadParameters = { pageNumber: 1, sortBy: 'created_at', sortDirection: 'desc' };
-
-export const afterSuccess = (dispatch) => {
-    dispatch(clearAddSource());
-    dispatch(loadEntities(afterSuccessLoadParameters));
-};
-
-export const prepareSourceTypeSelection = (sourceTypes) =>
-    sourceTypes.map(({ id, product_name }) => ({ label: product_name, value: id }));
-
-export const setFilter = (column, value, dispatch) => dispatch(filterProviders({
-    [column]: value
-}));
-
-export const chipsFormatters = (key, filterValue, sourceTypes) => ({
-    name: () => ({ name: filterValue[key], key }),
-    source_type_id: () => ({
-        category: 'Source Type',
-        key,
-        chips: filterValue[key].map(id => {
-            const sourceType = sourceTypes.find((type) => type.id === id);
-
-            return ({ name: sourceType ? sourceType.product_name : id, value: id });
-        })
-    })
-}[key] || (() => ({ name: key })));
-
-export const prepareChips = (filterValue, sourceTypes) =>
-    Object.keys(filterValue)
-    .map((key) =>
-        filterValue[key] && filterValue[key].length > 0 ? chipsFormatters(key, filterValue, sourceTypes)() : undefined
-    )
-    .filter(Boolean);
-
-export const removeChips = (chips, filterValue, deleteAll) => {
-    if (deleteAll) {
-        return (
-            Object.keys(filterValue).reduce((acc, curr) => ({
-                ...acc,
-                [curr]: undefined
-            }), {})
-        );
-    }
-
-    const chip = chips[0];
-
-    return ({
-        ...filterValue,
-        [chip.key]: chip.chips ? filterValue[chip.key].filter((value) => value !== chip.chips[0].value) : undefined
-    });
-};
+import {
+    prepareChips,
+    removeChips,
+    setFilter,
+    debouncedFiltering,
+    prepareSourceTypeSelection,
+    afterSuccess,
+    onCloseAddSourceWizard
+} from './SourcesPage/helpers';
 
 const SourcesPage = () => {
     const [showEmptyState, setShowEmptyState] = useState(false);
@@ -204,7 +134,7 @@ const SourcesPage = () => {
                 activeFiltersConfig={{
                     filters: prepareChips(filterValue, sourceTypes),
                     onDelete: (_event, chips, deleteAll) =>
-                        dispatch(filterProviders(removeChips(chips, filterValue, deleteAll)))
+                        dispatch(filterProviders(removeChips(chips, filterValue, deleteAll, setFilterValue)))
                 }}
             />
             <SourcesSimpleView />
